requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
langgraph>=0.2.60
httpx>=0.28.1
openai>=1.51.0
google-genai>=1.5.0
python-dotenv>=1.0.1
python-multipart>=0.0.12
